realestate/init/src/main/java/cn/gtmap/realestate/init/service/qlxx/qlfl/impl/InitLpbToBdcTdcbnydsyqServiceImpl.java
package cn.gtmap.realestate.init.service.qlxx.qlfl.impl;

import cn.gtmap.realestate.common.core.domain.BdcQl;
import cn.gtmap.realestate.common.core.domain.BdcTdcbnydsyqDO;
import cn.gtmap.realestate.init.core.qo.InitServiceQO;
import cn.gtmap.realestate.init.service.qlxx.qlfl.InitBdcTdcbnydsyqAbstractService;
import cn.gtmap.realestate.init.util.Constants;
import org.springframework.stereotype.Service;

/**
 * @author <a href="mailto:lisongtao@gtmap.cn">lisongtao</a>
 * @version 1.0  2018/11/6.
 * @description   从楼盘表加载数据到土地承包经营权、农用地使用权
 */
@Service
public class InitLpbToBdcTdcbnydsyqServiceImpl extends InitBdcTdcbnydsyqAbstractService {


    @Override
    public String getVal() {
        return Constants.QLSJLY_LPB;
    }


    @Override
    public BdcQl initQlxx(InitServiceQO initServiceQO) throws InstantiationException, IllegalAccessException {
        return initTdFromLpb(initServiceQO,BdcTdcbnydsyqDO.class);
    }
}
